move common elements of DownsampleVideoSchema to base schema
